+++ database/migrations/2023_03_02_220256_update_gitlab_merge_request_table_for_optional_fields.php
@@ -1,0 +1,36 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class UpdateGitlabMergeRequestTableForOptionalFields extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up(): void
+    {
+        Schema::table('gitlab_merge_requests', static function (Blueprint $table) {
+            $table->unsignedBigInteger('author_id')->nullable()->change();
+            $table->dateTimeTz('merge_request_created_at')->nullable()->change();
+            $table->dateTimeTz('merge_request_updated_at')->nullable()->change();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down(): void
+    {
+        Schema::table('gitlab_merge_requests', static function (Blueprint $table) {
+            $table->unsignedBigInteger('author_id')->change();
+            $table->dateTimeTz('merge_request_created_at')->change();
+            $table->dateTimeTz('merge_request_updated_at')->change();
+        });
+    }
+}