+++ src/Repositories/MergeRequestRepository.php
@@ -14,15 +14,23 @@ use TJVB\GitlabModelsForLaravel\Models\MergeRequest;
 
 final class MergeRequestRepository implements MergeRequestWriteRepository
 {
+    private array $optionalFields = [
+        'author_id',
+        'blocking_discussions_resolved',
+        'description',
+        'merge_request_created_at',
+        'work_in_progress',
+    ];
+
     public function updateOrCreate(int $mergeRequestId, array $mergeRequestData): MergeRequestContract
     {
-        return MergeRequest::updateOrCreate(['merge_request_id' => $mergeRequestId], [
-            'author_id' => (int) Arr::get($mergeRequestData, 'author_id'),
-            'blocking_discussions_resolved' => Arr::get($mergeRequestData, 'blocking_discussions_resolved', true),
-            'description' => (string) Arr::get($mergeRequestData, 'description'),
+        $updateData = [
+            'author_id' => Arr::get($mergeRequestData, 'author_id'),
+            'blocking_discussions_resolved' => Arr::get($mergeRequestData, 'blocking_discussions_resolved'),
+            'description' => Arr::get($mergeRequestData, 'description'),
             'merge_request_created_at' => CarbonImmutable::make(Arr::get($mergeRequestData, 'created_at')),
             'merge_request_iid' => (int) Arr::get($mergeRequestData, 'iid'),
-            'merge_status' => (string) Arr::get($mergeRequestData, 'merge_status'),
+            'merge_status' => Arr::get($mergeRequestData, 'merge_status'),
             'merge_request_updated_at' => CarbonImmutable::make(Arr::get($mergeRequestData, 'updated_at')),
             'state' => (string) Arr::get($mergeRequestData, 'state'),
             'source_project_id' => (int) Arr::get($mergeRequestData, 'source_project_id'),
@@ -31,8 +39,14 @@ final class MergeRequestRepository implements MergeRequestWriteRepository
             'target_branch' => (string) Arr::get($mergeRequestData, 'target_branch'),
             'title' => (string) Arr::get($mergeRequestData, 'title'),
             'url' => (string) Arr::get($mergeRequestData, 'url'),
-            'work_in_progress' => Arr::get($mergeRequestData, 'work_in_progress', false),
-        ]);
+            'work_in_progress' => Arr::get($mergeRequestData, 'work_in_progress'),
+        ];
+        foreach ($this->optionalFields as $field) {
+            if ($updateData[$field] === null) {
+                unset($updateData[$field]);
+            }
+        }
+        return MergeRequest::updateOrCreate(['merge_request_id' => $mergeRequestId], $updateData);
     }
 
     public function syncLabels(int $mergeRequestId, array $labels): ?MergeRequestContract